requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
python-multipart
pypdf
docx2txt
python-dotenv